% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_citation.R
\name{insert_citation}
\alias{insert_citation}
\title{Invoke RStudio addin to insert Markdown citations}
\usage{
insert_citation(bib_file = options("bibliography_path"))
}
\arguments{
\item{bib_file}{Character. Path to BibTeX-file. See details.}
}
\value{
Inserts selected Markdown citation(s) at currenct location.
}
\description{
Invoke RStudio addin to insert Markdown citations
}
\details{
The path to the BibTeX-file can be set in the global options and is set to
   \code{references.bib} when the package is loaded. Once the path is changed in the
   RStudio addin, the global option is updated.
}
\examples{
\dontrun{
 insert_citation(bib_file = "references.bib")
}
}
\seealso{
\code{\link{insert_citation}}
}

