% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_bib.R
\name{load_betterbiblatex_bib}
\alias{load_betterbiblatex_bib}
\title{Load bibliography from Zotero}
\usage{
load_betterbiblatex_bib(encoding, betterbiblatex_format = "bibtex",
  exclude_betterbiblatex_library = NULL, increment_progress = FALSE)
}
\description{
Load bibliography from Zotero
}
\examples{
\dontrun{
b <- load_betterbiblatex_bib(encoding = "UTF-8")
}
}
