% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_cite.R
\name{md_cite}
\alias{md_cite}
\title{Insert citation in Markdown format}
\usage{
md_cite(x, in_paren = TRUE, bib_file = getOption("citr.bibliography_path"),
  cache = TRUE, use_betterbiblatex = getOption("citr.use_betterbiblatex"))
}
\arguments{
\item{x}{Character. Search terms used to filter bibliography (by author, year, title, and journal
fields); Regex is supported.}

\item{in_paren}{Logical. Determines if citation is in parentheses.}

\item{bib_file}{Character. Path to Bib(La)TeX-file. See details.}

\item{cache}{Logical. If \code{cache = TRUE} cached bibliography is used, if available. If
\code{cache = FALSE} bibliography is re-imported on every function call.}

\item{use_betterbiblatex}{Logical. If \code{use_betterbiblatex = TRUE} references are imported from Zotero/Juris-M.
Requires that the \href{https://github.com/retorquere/zotero-better-bibtex}{Better Bib(La)TeX} is installed and
Zotero/Juris-M is running.}
}
\value{
If the bibliography contains exactly one match the formated citation is returned, otherwise
   returns \code{NULL}. \code{md_cite} returns an in-text citation (\code{"@foo2016"}), \code{md_cite}
   returns an in-parenthesis citation (\code{"[@foo2016]"}).
}
\description{
Look up entries in bibliography and insert citation in Markdown format if exactly one match is returned.
}
\details{
The path to the BibTeX-file can be set in the global options and is set to
   \code{references.bib} when the package is loaded. Once the path is changed in the
   RStudio addin, the global option is updated. If \code{use_betterbiblatex = TRUE} references
   are imported from Zotero/Juris-M rather than from the Bib(La)TeX-file. The Bib(La)TeX-file
   is then updated to include the inserted reference.
}
\examples{
\dontrun{
  md_cite("foo 2016", bib_file = "references.bib")
}

}
\seealso{
\code{\link{insert_citation}}
}

