% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_cite_zotero.R
\name{md_cite_zotero}
\alias{md_cite_zotero}
\title{Insert citation in Markdown format from Zotero}
\usage{
md_cite_zotero(in_paren = TRUE,
  bib_file = options("citr.bibliography_path"))
}
\arguments{
\item{in_paren}{Logical. Determines if citation is in parentheses.}

\item{bib_file}{Character. Path to Bib(La)TeX-file. See details.}
}
\value{
If the bibliography contains exactly one match the formated citation is returned, otherwise
   returns \code{NULL}. \code{md_cite} returns an in-text citation (\code{"@foo2016"}), \code{md_cite}
   returns an in-parenthesis citation (\code{"[@foo2016]"}).
}
\description{
Look up entries in Zotero database and insert citation in Markdown format.
}
\details{
The path to the Bib(La)TeX-file can be set in the global options and is set to
   \code{references.bib} when the package is loaded. Once the path is changed in the
   RStudio addin, the global option is updated.
}
\examples{
\dontrun{
  md_cite_zotero(bib_file = "references.bib")
}

}
\seealso{
\code{\link{insert_citation}}
}

