% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_bib_file.R
\name{tidy_bib_file}
\alias{tidy_bib_file}
\title{Tidy bibliography file}
\usage{
tidy_bib_file(
  rmd_file,
  messy_bibliography,
  file = NULL,
  keep_existing = "none",
  encoding = getOption("citr.encoding"),
  betterbiblatex_format = getOption("citr.betterbiblatex_format")
)
}
\arguments{
\item{rmd_file}{Character. One path or a vector of paths to the R Markdown files that use the messy bibliography file.}

\item{messy_bibliography}{Character. One path or a vector of paths to the messy bibliography file(s).}

\item{file}{Character. Path and name for the to-be-created tidy bibliography. If \code{NULL} the messy bibliography is replaced.}

\item{keep_existing}{Character. If \code{file} is not \code{NULL}, \code{keep_existing} determines which entries are kept. If \code{"all"}, all entries are kept. If \code{"none"}, all entries are replaced. If \code{"necessary"}, only entries that are in the R Markdown files are kept.}

\item{encoding}{Character. Encoding of the Bib(La)TeX-file. See the \code{\link{connection}}.}

\item{betterbiblatex_format}{Character. Bibliography format to export from Zotero/Juris-M. Can be either \code{"bibtex"} or \code{"biblatex"}. Ignored if \code{use_betterbiblatex = FALSE}.
Requires that the \href{https://github.com/retorquere/zotero-better-bibtex}{Better BibTeX} is installed and
Zotero/Juris-M is running.}
}
\description{
Removes duplicate and unneeded entries from a Bib(La)Tex-file.
}
\examples{
\dontrun{tidy_bib_file(rmd_file = c("introduction.Rmd", "methods.Rmd"), messy_bibliography = "references.bib", file = "tidy_references.bib")}
}
