% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_bib.R
\name{query_bib}
\alias{query_bib}
\title{Query bibliography}
\usage{
query_bib(x, bib_file = options("citr.bibliography_path"), cache = TRUE)
}
\arguments{
\item{x}{Character. Search terms used to filter bibliography (by author, year, title, and journal
fields); Regex is supported.}

\item{bib_file}{Character. Path to Bib(La)TeX-file. See details.}

\item{cache}{Logical. Cached bibliography is used, if available. If \code{cache = FALSE} bibliography
is re-imported on every function call.}
}
\value{
Returns list of class \code{\link[RefManageR]{BibEntry}} including all matching bibliography entries.
}
\description{
Search entries in a bibliography.
}
\details{
The path to the BibTeX-file can be set in the global options and is set to
   \code{references.bib} when the package is loaded. Once the path is changed in the
   RStudio addin, the global option is updated.
}
\seealso{
\code{\link{md_cite}}, \code{\link{insert_citation}}
}

