% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_bib_file.R
\name{tidy_bib_file}
\alias{tidy_bib_file}
\title{Tidy bibliography file}
\usage{
tidy_bib_file(rmd_file, messy_bibliography, file = NULL,
  encoding = getOption("citr.encoding"))
}
\arguments{
\item{rmd_file}{Character. One or more paths to the R Markdown files that use the messy bibliography file.}

\item{messy_bibliography}{Character. Path to the messy bibliography file.}

\item{file}{Character. Path and name for the to-be-created tidy bibliography. If \code{NULL} the messy bibliography is replaced.}

\item{encoding}{Character. The name of the encoding to be assumed. See the \code{\link{connection}}.}
}
\description{
Removes duplicate and unneeded entries from a Bib(La)Tex-file.
}
\examples{
NULL
}
